% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_users_input.R
\name{check_users_input}
\alias{check_users_input}
\title{Title Check user input to make sure it is one of a vector of acceptable answers, or 'Q' to quit.}
\usage{
check_users_input(acceptable_answers, user_input = NULL)
}
\arguments{
\item{acceptable_answers}{vector of acceptable answers, e.g. c("Y","N")}

\item{user_input}{Don't add anything to this, for internal function use.}
}
\value{
'Y', 'N' or break from function.
}
\description{
Title Check user input to make sure it is one of a vector of acceptable answers, or 'Q' to quit.
}
\examples{
\dontrun
}
