% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_imdp_data_prep.R
\name{figure_imdp_data_prep}
\alias{figure_imdp_data_prep}
\title{More stringent data preparation for IMDP report figures}
\usage{
figure_imdp_data_prep(
  dat = NULL,
  options_filepath = NA,
  year = NULL,
  verbose = T
)
}
\arguments{
\item{dat}{This should be the cleaned data output from the 'general_imdp_data_prep' function; i.e. a file named 'WatercraftInspectionData_AllYears_Selected_Columns.xlsx'. If NULL, will search J: LAN folder for data file.}

\item{options_filepath}{Filepath to 'my_opts.csv' file on local machine.}

\item{year}{Which year's IMDP final report are we prepping this data for? If NULL, will default to the year in 'my_opts.csv'}

\item{verbose}{Return copious feedback?}
}
\value{
Four excel files (this year's inspections, all inspections, high-risk inspections, and mussel-fouled inspections)
}
\description{
More stringent data preparation for IMDP report figures
}
\examples{
\dontrun
}
