% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_imdp_data_prep.R
\name{general_imdp_data_prep}
\alias{general_imdp_data_prep}
\title{Combine and clean IMDP inspection data}
\usage{
general_imdp_data_prep(options_filepath = NA, verbose = T)
}
\arguments{
\item{options_filepath}{Filepath to options .csv file}

\item{verbose}{Return copious feedback?}
}
\value{
Three excel files will be saved to the external output folder specified in the options file.
}
\description{
This is one of the main functions exported by the {imdp} package.
It collates and cleans excel and metabase data files. Adjust the data imports
and destination folders for the cleaned data files with the options .csv file.
}
\examples{
\dontrun
}
