% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all_imdp_data.R
\name{update_all_imdp_data}
\alias{update_all_imdp_data}
\title{Title Run both IMDP data cleaning functions in succession}
\usage{
update_all_imdp_data(
  dat = NULL,
  year = NULL,
  options_filepath =
    "C:/Users/CMADSEN/Downloads/LocalR/long_term_projects/ZQMussels/Options.csv",
  verbose = T,
  update_spatial_files = T
)
}
\arguments{
\item{dat}{This should be the cleaned data output from the 'general_imdp_data_prep' function; i.e. a file named 'WatercraftInspectionData_AllYears_Selected_Columns.xlsx'. If NULL, will search J: LAN folder for data file.}

\item{year}{Which year's IMDP final report are we prepping this data for? If NULL, will default to the year in 'my_opts.csv'}

\item{options_filepath}{Filepath to 'my_opts.csv' file on local machine.}

\item{verbose}{Return copious feedback?}
}
\value{
Four excel files (this year's inspections, all inspections, high-risk inspections, and mussel-fouled inspections), plus updated GIS files on W: drive and excel tables on LAN J: drive.
}
\description{
Title Run both IMDP data cleaning functions in succession
}
\examples{
\dontrun
}
