% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_old_data_watercraft_type.R
\name{update_old_data_watercraft_type}
\alias{update_old_data_watercraft_type}
\title{Title Update the older data's watercraft type.}
\usage{
update_old_data_watercraft_type(dat)
}
\arguments{
\item{dat}{table of IMDP inspections}
}
\value{
table of IMDP inspections, old records with watercraft complexity types
}
\description{
Title Update the older data's watercraft type.
}
\examples{
\dontrun
}
