% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_imdp_data_to_waterbodies.R
\name{summarise_imdp_data_to_waterbodies}
\alias{summarise_imdp_data_to_waterbodies}
\title{Title Perform the long-winded process of joining IMDP inspection data to waterbodies.}
\usage{
summarise_imdp_data_to_waterbodies(
  dat = NULL,
  options_filepath =
    "C:/Users/CMADSEN/Downloads/LocalR/long_term_projects/ZQMussels/Options.csv",
  redo_geocoding = F,
  verbose = T
)
}
\arguments{
\item{dat}{This should be the cleaned data output from the 'general_imdp_data_prep' function; i.e. a file named 'WatercraftInspectionData_AllYears_Selected_Columns.xlsx'. If NULL, will search J: LAN folder for data file.}

\item{options_filepath}{Filepath to 'my_opts.csv' file on local machine.}

\item{redo_geocoding}{Shall we redo the time-intensive geocoding of closest cities? This takes about 20 minutes.}

\item{verbose}{Return copious feedback?}
}
\value{
Four excel files (this year's inspections, all inspections, high-risk inspections, and mussel-fouled inspections)
}
\description{
Title Perform the long-winded process of joining IMDP inspection data to waterbodies.
}
\examples{
\dontrun
}
